requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
frozenlist==1.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
oauthlib==3.1.1
pyee==8.2.2
python-dotenv==0.19.1
requests==2.26.0
requests-oauthlib==1.3.0
slackclient==2.9.3
slackeventsapi==3.0.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
